Adds the beginings of a LocationInfo class
I'm moving to a better design philosophy, to make this project less messy.
LocationInfo is the class containing info about a Location, which is was called a relPath. It is the relative path from any particular root.
So, if you are joining roots /usr/me/ and /usr/you/, and me has a file hi.txt in it, then /hi.txt would be that file's location.
They are important because essentially all the joiner will do is watch all files with the same location, and ensure that they stay synced.