Add revamped classes to track the state of a file system tree, and a logical aggregate tree
I'm calling a "path" a particular path on the filesystem, and a "location" a relative location from a root.
This adds:
PathInfo: watches a particular path for changes to its file contents, including its child files if a directory.
PathTree: the summary object for a tree of paths, sprouting from a given root. This can be watched for info about what changes occur in the tree.

LocationInfo: watches all the PathInfo's at a particular location over a number of PathTree's
LocationTree: thge summary object for a logical tree of LocationInfo's. This can be watched for info about canges to the "logical aggregate tree"

The aggregate tree has a kind of union over the paths of all the passed in PathTrees, with precedence to those with lower indexes in the passed in array